+++ srcs/equationpoly4.c
@@ -24,19 +24,19 @@ static t_res	fin(double u, double aa, double bb, double bs)
 	res.x4 = NAN;
 	res.nb = 0;
 	uma = u - aa;
-	d = uma - 4.0 * (bb / 2 / uma * sqrt(uma) + u / 2);
+	d = uma - (double)4.0 * (bb / (double)2 / uma * sqrt(uma) + u / (double)2);
 	if (d >= 0)
 	{
 		res.nb += 2;
-		res.x1 = (sqrt(uma) + sqrt(d)) / 2 - bs;
-		res.x2 = (sqrt(uma) - sqrt(d)) / 2 - bs;
+		res.x1 = (double)(sqrt(uma) + sqrt(d)) / (double)2 - bs;
+		res.x2 = (double)(sqrt(uma) - sqrt(d)) / (double)2 - bs;
 	}
-	d = uma - 4.0 * (-bb / 2 / uma * sqrt(uma) + u / 2);
+	d = uma - (double)4.0 * (-bb / (double)2 / uma * sqrt(uma) + u / (double)2);
 	if (d >= 0)
 	{
 		res.nb += 2;
-		res.x3 = (-sqrt(uma) + sqrt(d)) / 2 - bs;
-		res.x4 = (-sqrt(uma) - sqrt(d)) / 2 - bs;
+		res.x3 = (double)(-sqrt(uma) + sqrt(d)) / (double)2 - bs;
+		res.x4 = (double)(-sqrt(uma) - sqrt(d)) / (double)2 - bs;
 	}
 	return (res);
 }
@@ -45,13 +45,13 @@ static t_res	delta_negative(t_equ var, t_res res)
 {
 	if (var.p != 0)
 	{
-		var.kos = -var.q / 2 / sqrt(-var.p * var.p * var.p / 27);
-		var.r = sqrt(-var.p / 3);
+		var.kos = -var.q / (double)2 / sqrt(-var.p * var.p * var.p / (double)27);
+		var.r = sqrt(-var.p / (double)3);
 	}
 	var.alpha = acos(var.kos);
-	res.x1 = 2 * var.r * cos((var.alpha) / 3) + var.vt;
-	res.x2 = 2 * var.r * cos((var.alpha + 2 * M_PI) / 3) + var.vt;
-	res.x3 = 2 * var.r * cos((var.alpha + 2 * 2 * M_PI) / 3) + var.vt;
+	res.x1 = (double)2 * var.r * cos((var.alpha) / (double)3) + var.vt;
+	res.x2 = (double)2 * var.r * cos((var.alpha + (double)2 * M_PI) / (double)3) + var.vt;
+	res.x3 = (double)2 * var.r * cos((var.alpha + (double)2 * (double)2 * M_PI) / (double)3) + var.vt;
 	if (var.r == 0)
 		return (fin(res.x1, var.aa, var.bb, var.bs));
 	else if (res.x1 > var.aa)
@@ -69,32 +69,32 @@ static void		var_init(t_equ *var, t_res *res, double *val)
 	res->x3 = NAN;
 	res->x4 = NAN;
 	res->nb = 0;
-	var->bs = val[1] / 4 / val[0];
-	var->aa = -3 * val[1] * val[1] / 8 / val[0] / val[0] + val[2] / val[0];
+	var->bs = val[1] / (double)4 / val[0];
+	var->aa = -(double)3 * val[1] * val[1] / 8 / val[0] / val[0] + val[2] / val[0];
 	var->ma = -var->aa;
 	var->bb = val[1] * val[1] * val[1] / 8 / val[0] / val[0] / val[0] - val[1]
 	* val[2] / 2 / val[0] / val[0] + val[3] / val[0];
-	var->cc = -3 * val[1] * val[1] * val[1] * val[1] / 256 / val[0] / val[0]
-	/ val[0] / val[0] + val[2] * val[1] * val[1] / 16 / val[0] / val[0]
-	/ val[0] - val[1] * val[3] / 4 / val[0] / val[0] + val[4] / val[0];
-	var->mc = -4 * var->cc;
-	var->k = 4 * var->aa * var->cc - var->bb * var->bb;
-	var->kos = 0;
-	var->r = 0;
+	var->cc = -(double)3 * val[1] * val[1] * val[1] * val[1] / (double)256 / val[0] / val[0]
+	/ val[0] / val[0] + val[2] * val[1] * val[1] / (double)16 / val[0] / val[0]
+	/ val[0] - val[1] * val[3] / (double)4 / val[0] / val[0] + val[4] / val[0];
+	var->mc = -(double)4 * var->cc;
+	var->k = (double)4 * var->aa * var->cc - var->bb * var->bb;
+	var->kos = (double)0;
+	var->r = (double)0;
 }
 
 static void		var_calcul(t_equ *var, double *val)
 {
-	var->bs = val[1] / 4 / val[0];
-	val[0] = 1;
+	var->bs = val[1] / (double)4 / val[0];
+	val[0] = (double)1;
 	val[1] = -var->aa;
 	val[2] = var->mc;
 	val[3] = var->k;
 	var->vt = -val[1] / 3 / val[0];
-	var->p = val[2] / val[0] - val[1] * val[1] / 3 / val[0] / val[0];
-	var->q = val[1] * val[1] * val[1] / val[0] / val[0] / val[0] / 13.5
-	+ val[3] / val[0] - val[1] * val[2] / 3 / val[0] / val[0];
-	var->del = var->q * var->q / 4 + var->p * var->p * var->p / 27;
+	var->p = val[2] / val[0] - val[1] * val[1] / (double)3 / val[0] / val[0];
+	var->q = val[1] * val[1] * val[1] / val[0] / val[0] / val[0] / (double)13.5
+	+ val[3] / val[0] - val[1] * val[2] / (double)3 / val[0] / val[0];
+	var->del = var->q * var->q / (double)4 + var->p * var->p * var->p / (double)27;
 }
 
 t_res			equationpoly4(double *val)
